requirements.collector.txt
kafka-python==2.0.2
pandas==1.5.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
slack_sdk>=3.0